requirements.txt
streamlit==1.31.0
pdfplumber==0.10.3
nltk==3.8.1
torch==2.7.0
sentence-transformers==2.2.2
huggingface-hub==0.16.4
scikit-learn==1.3.2
pandas==2.2.0
matplotlib==3.8.2
plotly==5.18.0
openai==1.12.0
python-dotenv==1.0.1